move april module
create aruco_utils.hpp

move Board, GridBoard, CharucoBoard to board.hpp/board.cpp

refactoring _getSingleMarkerObjectPoints()

refactoring _extractBits()

refactoring _findMarkerContours()

fix _copyVector2Output() in detectMarkers()

move testCharucoCornersCollinear() to board.hpp/board.cpp

move poseEstimate()/calibAruco() to aruco_calib_pose.hpp

reduce include files

move detectMarkers() to class ArucoDetector

move refineDetectedMarkers() to class ArucoDetector

add C API wrapper to detectMarkers(), refineDetectedMarkers()

update tests and samples to class API

add py tests: test_aruco_detector, test_aruco_detector_refine

refactoring, fix docs

add java tests: testArucoIssue3133, testArucoDetector

add readWriteParameter(), update readParameter()

implemented cv::Algorithm - read/write, added read/write to RefineParameters, added write to DetectorParameters

merge PatternPos/EstimateParameters after rebase

remove empty docstring for private function

fixes

fixes license
