+++ modules/aruco/include/opencv2/aruco/dictionary.hpp
@@ -1,41 +1,6 @@
-/*
-By downloading, copying, installing or using the software you agree to this
-license. If you do not agree to this license, do not download, install,
-copy or use the software.
-
-                          License Agreement
-               For Open Source Computer Vision Library
-                       (3-clause BSD License)
-
-Copyright (C) 2013, OpenCV Foundation, all rights reserved.
-Third party copyrights are property of their respective owners.
-
-Redistribution and use in source and binary forms, with or without modification,
-are permitted provided that the following conditions are met:
-
-  * Redistributions of source code must retain the above copyright notice,
-    this list of conditions and the following disclaimer.
-
-  * Redistributions in binary form must reproduce the above copyright notice,
-    this list of conditions and the following disclaimer in the documentation
-    and/or other materials provided with the distribution.
-
-  * Neither the names of the copyright holders nor the names of the contributors
-    may be used to endorse or promote products derived from this software
-    without specific prior written permission.
-
-This software is provided by the copyright holders and contributors "as is" and
-any express or implied warranties, including, but not limited to, the implied
-warranties of merchantability and fitness for a particular purpose are
-disclaimed. In no event shall copyright holders or contributors be liable for
-any direct, indirect, incidental, special, exemplary, or consequential damages
-(including, but not limited to, procurement of substitute goods or services;
-loss of use, data, or profits; or business interruption) however caused
-and on any theory of liability, whether in contract, strict liability,
-or tort (including negligence or otherwise) arising in any way out of
-the use of this software, even if advised of the possibility of such damage.
-*/
-
+// This file is part of OpenCV project.
+// It is subject to the license terms in the LICENSE file found in the top-level directory
+// of this distribution and at http://opencv.org/license.html
 #ifndef __OPENCV_DICTIONARY_HPP__
 #define __OPENCV_DICTIONARY_HPP__
 
@@ -66,28 +31,22 @@ class CV_EXPORTS_W Dictionary {
     CV_PROP_RW int maxCorrectionBits; // maximum number of bits that can be corrected
 
 
-    /**
-      */
     Dictionary(const Mat &_bytesList = Mat(), int _markerSize = 0, int _maxcorr = 0);
 
 
     /**
-    Dictionary(const Dictionary &_dictionary);
+    * Dictionary(const Dictionary &_dictionary);
     */
+    Dictionary(const Ptr<Dictionary> &dictionary);
 
 
-    /**
-      */
-    Dictionary(const Ptr<Dictionary> &_dictionary);
-
-
-    /**
+    /** @brief returns generateCustomDictionary(nMarkers, markerSize, randomSeed)
      * @see generateCustomDictionary
      */
     CV_WRAP_AS(create) static Ptr<Dictionary> create(int nMarkers, int markerSize, int randomSeed=0);
 
 
-    /**
+    /** @brief returns generateCustomDictionary(nMarkers, markerSize, baseDictionary, randomSeed)
      * @see generateCustomDictionary
      */
     CV_WRAP_AS(create_from) static Ptr<Dictionary> create(int nMarkers, int markerSize,
@@ -108,6 +67,7 @@ class CV_EXPORTS_W Dictionary {
      * @brief Write a dictionary to FileStorage. Format is the same as in readDictionary().
      */
     CV_WRAP void writeDictionary(Ptr<FileStorage>& fs);
+
     /**
      * @see getPredefinedDictionary
      */
