+++ modules/aruco/perf/perf_aruco.cpp
@@ -190,6 +190,7 @@ PERF_TEST_P(EstimateAruco, ArucoFirst, ESTIMATE_PARAMS)
         detectorParams->minSideLengthCanonicalImg = 32;
         detectorParams->minMarkerLengthRatioOriginalImg = 0.04f / numMarkersInRow;
     }
+    aruco::ArucoDetector detector(dictionary, detectorParams);
     MarkerPainter painter(markerSize);
     auto image_map = painter.getProjectMarkersTile(numMarkersInRow, detectorParams, dictionary);
 
@@ -198,7 +199,7 @@ PERF_TEST_P(EstimateAruco, ArucoFirst, ESTIMATE_PARAMS)
     vector<int> ids;
     TEST_CYCLE()
     {
-        aruco::detectMarkers(image_map.first, dictionary, corners, ids, detectorParams);
+        detector.detectMarkers(image_map.first, corners, ids);
     }
     ASSERT_EQ(numMarkersInRow*numMarkersInRow, static_cast<int>(ids.size()));
     double maxDistance = getMaxDistance(image_map.second, ids, corners);
@@ -221,6 +222,7 @@ PERF_TEST_P(EstimateAruco, ArucoSecond, ESTIMATE_PARAMS)
         detectorParams->minSideLengthCanonicalImg = 64;
         detectorParams->minMarkerLengthRatioOriginalImg = 0.f;
     }
+    aruco::ArucoDetector detector(dictionary, detectorParams);
     const int markerSize = 200;
     const int numMarkersInRow = 11;
     MarkerPainter painter(markerSize);
@@ -231,7 +233,7 @@ PERF_TEST_P(EstimateAruco, ArucoSecond, ESTIMATE_PARAMS)
     vector<int> ids;
     TEST_CYCLE()
     {
-        aruco::detectMarkers(image_map.first, dictionary, corners, ids, detectorParams);
+        detector.detectMarkers(image_map.first, corners, ids);
     }
     ASSERT_EQ(numMarkersInRow*numMarkersInRow, static_cast<int>(ids.size()));
     double maxDistance = getMaxDistance(image_map.second, ids, corners);
@@ -276,6 +278,7 @@ PERF_TEST_P(EstimateLargeAruco, ArucoFHD, ESTIMATE_FHD_PARAMS)
         detectorParams->minSideLengthCanonicalImg = get<0>(testParams).minSideLengthCanonicalImg;
         detectorParams->minMarkerLengthRatioOriginalImg = get<0>(testParams).minMarkerLengthRatioOriginalImg;
     }
+    aruco::ArucoDetector detector(dictionary, detectorParams);
     const int markerSize = get<1>(testParams).first;       // 1440 or 480 or 144
     const int numMarkersInRow = get<1>(testParams).second; // 1 or 3 or 144
     MarkerPainter painter(markerSize);                     // num pixels is 1440x1440 as in FHD 1920x1080
@@ -286,7 +289,7 @@ PERF_TEST_P(EstimateLargeAruco, ArucoFHD, ESTIMATE_FHD_PARAMS)
     vector<int> ids;
     TEST_CYCLE()
     {
-        aruco::detectMarkers(image_map.first, dictionary, corners, ids, detectorParams);
+        detector.detectMarkers(image_map.first, corners, ids);
     }
     ASSERT_EQ(numMarkersInRow*numMarkersInRow, static_cast<int>(ids.size()));
     double maxDistance = getMaxDistance(image_map.second, ids, corners);
