+++ modules/aruco/samples/aruco_samples_utility.hpp
@@ -1,5 +1,5 @@
 #include <opencv2/highgui.hpp>
-#include <opencv2/aruco.hpp>
+#include <opencv2/aruco_detector.hpp>
 #include <opencv2/calib3d.hpp>
 #include <ctime>
