+++ modules/aruco/samples/create_board.cpp
@@ -38,7 +38,7 @@ the use of this software, even if advised of the possibility of such damage.
 
 
 #include <opencv2/highgui.hpp>
-#include <opencv2/aruco.hpp>
+#include <opencv2/aruco_detector.hpp>
 #include <iostream>
 #include "aruco_samples_utility.hpp"
