+++ modules/aruco/src/apriltag/predefined_dictionaries_apriltag.hpp
