+++ modules/aruco/src/apriltag/zmaxheap.cpp
@@ -12,7 +12,7 @@
 // of the authors and should not be interpreted as representing official policies,
 // either expressed or implied, of the Regents of The University of Michigan.
 
-#include "precomp.hpp"
+#include "../precomp.hpp"
 #include "zmaxheap.hpp"
