+++ modules/aruco/test/test_precomp.hpp
@@ -7,7 +7,6 @@
 #include "opencv2/ts.hpp"
 #include "opencv2/imgproc.hpp"
 #include "opencv2/calib3d.hpp"
-#include "opencv2/aruco.hpp"
 #include <opencv2/aruco/charuco.hpp>
 
 #endif
